+++ tests/integration_tests/test_parameter_filter.py
@@ -3,12 +3,12 @@ import pandas as pd
 from webviz_subsurface._components.parameter_filter import ParameterFilter
 
 
-def test_dataframe(app, testdata_folder) -> None:
+def test_dataframe(testdata_folder) -> None:
 
     dframe = pd.read_csv(
         testdata_folder / "reek_test_data" / "aggregated_data" / "parameters.csv"
     )
-    component = ParameterFilter(app, "test", dframe)
+    component = ParameterFilter("test", dframe)
     # pylint: disable=protected-access
     assert set(component._discrete_parameters) == set(
         ["FWL", "MULTFLT_F1", "INTERPOLATE_WO", "COHIBA_MODEL_MODE", "RMS_SEED"]